backend/requirements.txt
flask
numpy
pandas
scikit-learn
ollama
langchain
langchain-huggingface
transformers
huggingface-hub
python-dotenv
scikit-learn
langchain_community
langserve
fastapi
langchainhub
sentence_transformers
langchain-objectbox
flask-cors
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client 